backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# Official MariaDB Connector/Python: https://pypi.org/project/mariadb/
mariadb>=1.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
